# Feature or Bug Fix
[Issue 3260602] SP connector batch authZ should handle site collection URLs correctly for authZ grouping of URLs to avoid intermittent indeterminate results.

The authorization failed for default.aspx page URL for a site, when the site collection for the web application was created using the HTTP port 80.

http://SharePointSiteURL/default.aspx

# High level list of changes

o Modified code for 'GetServerRelativeWebUrl' function and added code to retrieve correct server relative URL for a site.
o The logic works for the web application created using HTTP port 80 as well.

# Specific changes


GSBulkAuthorization.asmx
-----------------------
o Modified code in GetServerRelativeWebUrl function, for extracting server relative URL by checking whether the url contains port number or not. If the url contains port number, find the colon and accordingly extract substring ahead of the colon. If the url does not contain port number, find the position for forward slash (/) and  extract substring ahead of the two forward slashes.
o Logic works even when the web application is created using HTTP port 80.

o Tested the working of Authorize web method for following URL Patterns - 

*** URLs of web applications using Port No. in the path  ***

• URL of root site's landing page using "/sites/" path in site collection 
http://SharePointSiteURL:portno/sites/ntlm1/default.aspx

• Social Site using "/sites/" path in site collection 
http://SharePointSiteURL:portno/sites/ntlm1/socialsite/default.aspx

• Blog Site using "/sites/" path in site collection 
http://SharePointSiteURL:portno/sites/ntlm1/blogsite/default.aspx

• URL of root site's landing page using "/" path in site collection 
http://SharePointSiteURL:portno/default.aspx

• Social Site using "/" path in site collection 
http://SharePointSiteURL:portno/socialsite/default.aspx

• Blog Site using "/" path in site collection 
http://SharePointSiteURL:portno/blogsite/default.aspx




*** URLs of web applications not using Port No. in the path (i.e. using port as
80) ***
• URL of root site's landing page using "/sites/" path in site collection 
http://SharePointSiteURL/sites/mysite/default.aspx 

• Social Site using "/sites/" path in site collection 
http://SharePointSiteURL/sites/mysite/socialsite/default.aspx 

• Blog Site using "/sites/" path in site collection 
http://SharePointSiteURL/sites/mysite/blogsite/default.aspx 

• URL of root site's landing page using "/" path in site collection 
http://SharePointSiteURL/default.aspx 

• Social Site using "/" path in site collection 
http://SharePointSiteURL/socialsite/default.aspx 

• Blog Site using "/" path in site collection 
http://SharePointSiteURL/blogsite/default.aspx